requirements.txt
langchain-core
langchain-community
langchain-experimental
psycopg2-binary
sqlalchemy
python-dotenv
pyyaml
fastapi
uvicorn
pydantic
langchain
langchain-core
